feat(insights): fold problem.* router into insight.*
Extend insight.list/create/update with the general triage fields (impact,
confidence, category, evidence) and add park/unpark + an includeParked filter,
mirroring the retired problem router. Remove problemRouter, its registration, its
tests, and the Approach procedures (linkProject/unlinkProject/setProjects/
productProjects) — the Problem<->Project links are dropped (ADR-0036). Access
control unchanged (product->workspace scoping via loadInsightWithAccess).

Problem UI consumers (Problems tab/page, ProblemKanbanBoard) are folded into the
Insights surface in the following slices on this branch.

Refs: Clear #27
Exponential-Ticket: cmr24piwy000lic04s95txeiy
